requirements.txt
numpy
matplotlib
networkx
cvxpy
scikit-learn
scipy
pillow
dccp
jupyter
notebook